library building - add spark specific directory to unmanagedSources (depends on current spark version) Add datafarme encoding Move encoders to seperated package
