+++ configs/router-examples-spark1.conf
@@ -19,11 +19,6 @@ simple-context = {
   namespace = "foo"
 }
 
-simple-context2 = {
-  path = ${jar_path}
-  className = "SimpleContextV2$"
-  namespace = "foo"
-}
 
 simple-context2-java = {
   path = ${jar_path}
@@ -40,5 +35,19 @@ simple-context-py = {
 hive-job-py = {
   path = ${py_path}"/simple_hive_context.py"
   className = "SimpleHiveContext"
+  namespace = "foo"
+}
+
+
+
+simple-context2 = {
+  path = ${jar_path}
+  className = "SimpleContextV2$"
+  namespace = "foo"
+}
+
+hive-context2 = {
+  path = ${jar_path}
+  className = "SimpleHiveContextV2$"
   namespace = "foo"
 }
