--- mist-lib/src/main/scala-2.10/mist/api/SparkDataEncoders.scala
@@ -1,4 +1,0 @@
-package mist.api
-
-trait SparkDataEncoders
-object SparkDataEncoders extends SparkDataEncoders
