+++ mist-lib/src/main/scala/mist/api/encoding/DefaultEncoders.scala
@@ -1,0 +1,7 @@
+package mist.api.encoding
+
+trait DefaultEncoders extends PrimitiveEncoders
+  with CollectionsEncoder
+  with DataFrameEncoding
+
+object DefaultEncoders extends DefaultEncoders
