+++ mist-lib/src/main/scala/mist/api/encoding/Encoder.scala
@@ -1,0 +1,12 @@
+package mist.api.encoding
+
+import mist.api.data.JsLikeData
+
+import scala.annotation.implicitNotFound
+
+@implicitNotFound(msg = "Could not find Encoder instance for ${A}")
+trait Encoder[A] extends Serializable {
+
+  def apply(a : A): JsLikeData
+
+}
