+++ mist-lib/src/test/resources/hive_job_data.json
@@ -1,0 +1,5 @@
+[
+  { "name" : "x", "age": 10 },
+  { "name" : "y", "age": 20 },
+  { "name" : "z", "age": 30 }
+]
