+++ mist-lib/src/test/scala/mist/api/ArgDefSpec.scala
@@ -6,7 +6,6 @@ import shapeless.HNil
 
 class ArgDefSpec extends FunSpec with Matchers {
 
-  import DefaultEncoders._
   import JobDefInstances._
 
   describe("const/missing") {
