+++ mist-lib/src/test/scala/mist/api/BaseContextsSpec.scala
@@ -1,30 +1,14 @@
 package mist.api
 
 import io.hydrosphere.mist.api.{RuntimeJobInfo, SetupConfiguration}
-import org.apache.spark.{SparkConf, SparkContext}
+import org.apache.spark.SparkContext
+import org.apache.spark.sql.hive.HiveContext
 import org.scalatest.{BeforeAndAfterAll, FunSpec, Matchers}
 
 class BaseContextsSpec extends FunSpec with Matchers with BeforeAndAfterAll {
 
-  val conf = new SparkConf()
-    .setMaster("local[2]")
-    .setAppName("mist-lib-test")
-    .set("spark.driver.allowMultipleContexts", "true")
-    .set("spark.ui.disabled", "true")
-
-  var sc: SparkContext = _
-
-  override def beforeAll(): Unit = {
-    sc = new SparkContext(conf)
-  }
-
-  override def afterAll = {
-    sc.stop()
-  }
-
-  import JobDefInstances._
-  import DefaultEncoders._
   import BaseContexts._
+  import JobDefInstances._
 
   it("for spark context") {
    val spJob = arg[Seq[Int]]("nums").onSparkContext(
@@ -44,9 +28,25 @@ class BaseContextsSpec extends FunSpec with Matchers with BeforeAndAfterAll {
     res shouldBe JobSuccess(5)
   }
 
+  it("for only hive context") {
+    def pathToResource(path: String): String = {
+      this.getClass.getClassLoader.getResource(path).getPath
+    }
+
+    val spJob = onHiveContext((hiveCtx: HiveContext) => {
+      val df = hiveCtx.read.json(pathToResource("hive_job_data.json"))
+      df.registerTempTable("temp")
+      hiveCtx.sql("SELECT MAX(age) AS avg_age FROM temp")
+        .take(1)(0).getLong(0)
+    })
+
+    val res = spJob.invoke(testCtx())
+    res shouldBe JobSuccess(30)
+  }
+
   def testCtx(params: (String, Any)*): JobContext = {
     val duration = org.apache.spark.streaming.Duration(10 * 1000)
-    val setupConf = SetupConfiguration(sc, duration, RuntimeJobInfo("test", "worker"), None)
+    val setupConf = SetupConfiguration(TestSparkContext.sc, duration, RuntimeJobInfo("test", "worker"), None)
     JobContext(setupConf, params.toMap)
   }
 }
